+++ system/src/Grav/Common/Processors/PluginsProcessor.php
@@ -23,6 +23,8 @@ class PluginsProcessor extends ProcessorBase
         $this->startTimer();
         // TODO: remove in 2.0.
         $this->container['accounts'];
+        // TODO: remove in 2.0.
+        $this->container['pages'];
         $this->container['plugins']->init();
         $this->container->fireEvent('onPluginsInitialized');
         $this->stopTimer();